Change dB value in trace to integer.

+++ examples/avr-rss2/ipv6/mqtt-noisesensor/mqtt-noisesensor.c
@@ -745,7 +745,7 @@ publish_sensors(void)
   /* Use device URN as base name -- draft-arkko-core-dev-urn-03 */  
   PUTFMT("{\"node_id\":\"%s\"", node_id);
   PUTFMT(",\"db\":%d}", value(0));
-  printf("printing publish_sensors: NODE_ID=%s dB=%-4.2f\n", node_id, ((double)value(0)));
+  printf("printing publish_sensors: NODE_ID=%s dB=%d\n", node_id, value(0));
 
   DBG("MQTT publish sensors %d: %d bytes\n", seq_nr_value, strlen(app_buffer));
   //printf("%s\n", app_buffer);
